tests: create mitoai testing structure
