+++ tests/jupyter_utils/mitosheet_utils.ts
@@ -1,4 +1,5 @@
 import { IJupyterLabPageFixture } from "@jupyterlab/galata";
+import { waitForIdle } from "./jupyterlab_utils";
 
 export const dfCreationCode = `import pandas as pd
 df = pd.DataFrame({'a': [1], 'b': [4]})\n`;
@@ -10,17 +11,6 @@ os.environ['MITO_CONFIG_DISABLE_TOURS'] = 'True'
 `
 
 type ToolbarButton = 'Insert' | 'Delete'
-
-export const createAndRunNotebookWithCells = async (page: IJupyterLabPageFixture, cellContents: string[]) => {
-  const randomFileName = `$test_file_${Math.random().toString(36).substring(2, 15)}.ipynb`;
-  await page.notebook.createNew(randomFileName);
-
-  for (let i = 0; i < cellContents.length; i++) {
-    await page.notebook.setCell(i, 'code', cellContents[i]);
-    await page.notebook.runCell(i);
-  }
-  await waitForIdle(page)
-}
 
 // If the test just interacts with the mitosheet, and not JupyterLab
 export const createNewMitosheetOnlyTest = async (page: IJupyterLabPageFixture, firstCellCode: string) => {
@@ -45,25 +35,6 @@ export const getNumberOfColumns = async (page: IJupyterLabPageFixture, cellNumbe
   return columns?.length || 0;
 }
 
-export const waitForIdle = async (page: IJupyterLabPageFixture) => {
-  const idleLocator = page.locator('#jp-main-statusbar >> text=Idle');
-  await idleLocator.waitFor();
-}
-
-export const waitForCodeToBeWritten = async (page: IJupyterLabPageFixture, cellIndex: number) => {
-  await waitForIdle(page);
-  const cellInput = await page.notebook.getCellInput(cellIndex);
-  let cellCode = (await cellInput?.innerText())?.trim();
-  // We wait until there's any code in the cell
-  while (!/[a-zA-Z]/g.test(cellCode || '')) {
-    // Wait 20 ms
-    await page.waitForTimeout(20);
-    await waitForIdle(page);
-    const cellInput = await page.notebook.getCellInput(cellIndex);
-    cellCode = (await cellInput?.innerText())?.trim();
-  }
-}
-
 export const updateCellValue = async (page: IJupyterLabPageFixture, cellValue: string, newCellValue: string) => {
   await page.locator('.mito-grid-cell', { hasText: cellValue }).scrollIntoViewIfNeeded();
   await page.locator('.mito-grid-cell', { hasText: cellValue }).dblclick();
@@ -71,9 +42,3 @@ export const updateCellValue = async (page: IJupyterLabPageFixture, cellValue: s
   await page.keyboard.press('Enter');
   await waitForIdle(page);
 };
-
-export const typeInNotebookCell = async (page: IJupyterLabPageFixture, cellIndex: number, cellValue: string) => {
-  await page.locator('.jp-Cell-inputArea').nth(cellIndex).scrollIntoViewIfNeeded();
-  await page.notebook.enterCellEditingMode(cellIndex);
-  await page.notebook.setCell(cellIndex, 'code', cellValue);
-}
